10.05

+++ src/main/java/com/wilczek/customer/CastomerDAO.java
@@ -7,15 +7,15 @@ import java.util.Optional;
 public interface CastomerDAO {
     List<Customer> selectAllCustomers();
 
-    Optional<Customer> selectCustomerById(Integer id);
+    Optional<Customer> selectCustomerById(Long id);
 
     void insertCustomer(Customer customer);
 
-    void deleteCustomerById(Integer id);
+    void deleteCustomerById(Long id);
 
     void updateCustomer(Customer customer);
 
     boolean existPersonWithEmail(String email);
 
-    boolean existPersonWithId(Integer id);
+    boolean existPersonWithId(Long id);
 }
